Refactor code to separate files + add dotenv config
