+++ amplify/.config/project-config.json
@@ -3,7 +3,7 @@
     "awscloudformation"
   ],
   "projectName": "YourClearWay",
-  "version": "3.1",
+  "version": "3.0",
   "frontend": "javascript",
   "javascript": {
     "framework": "react-native",
