+++ amplify/backend/function/questionslist/src/.env.example
@@ -1,0 +1,2 @@
+AWS_REGION=
+TABLE_NAME=
