+++ amplify/backend/function/questionslist/src/.gitignore
@@ -1,0 +1,1 @@
+.env
