+++ amplify/backend/function/questionslist/src/config.js
@@ -1,0 +1,10 @@
+const dotenv = require('dotenv');
+
+dotenv.config();
+
+const Config = {
+  AWS_REGION: process.env.AWS_REGION,
+  TABLE_NAME: process.env.TABLE_NAME,
+}
+
+module.exports = { Config };
