+++ amplify/backend/function/questionslist/src/event.json
@@ -1,5 +1,1 @@
-{
-  "key1": "value1",
-  "key2": "value2",
-  "key3": "value3"
-}
+{}
