+++ amplify/backend/function/questionslist/src/package.json
@@ -7,6 +7,7 @@
   "dependencies": {
     "aws-sdk": "^2.858.0",
     "csv-parse": "^4.15.3",
+    "dotenv": "^8.2.0",
     "uuid": "^8.3.2"
   }
 }
